Extract "bower-dependency-version-checker" file
